requirements.txt
reflex~=0.8.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
websockets>=12.0
pydantic>=2.0.0